#COBUMMFO-7736# Коментар: Merge.

+++ sql/mmfo/bars/Procedure/p_fe9_sb.sql
@@ -7,16 +7,20 @@ PROMPT =========================================================================
 
 PROMPT *** Create  procedure P_FE9_SB ***
 
-CREATE OR REPLACE PROCEDURE BARS.P_FE9_SB ( dat_     DATE,
-                                       sheme_   VARCHAR2 DEFAULT 'C') IS
+CREATE OR REPLACE PROCEDURE BARS.P_FE9_SB (
+                        dat_     DATE,
+                        sheme_   VARCHAR2 DEFAULT 'C',
+                        type_    varchar2 default ' ') IS
 /*%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 % DESCRIPTION : Процедура формирования #E9 для КБ
 % COPYRIGHT   : Copyright UNITY-BARS Limited, 1999.  All Rights Reserved.
-% VERSION     : 10/05/2018 (22/04/2018)
+% VERSION     : 21/05/2018      (10/05/2018, 22/04/2018)
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 параметры: Dat_ - отчетная дата
            sheme_ - схема формирования
+           type_  - ' '/'X' -обычный файл / подготовка xml
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+21.05.2018  для формирования xml -не сохранение в TMP_NBU
 05/04/2018 - для операций M37, MMV, CN3, CN4
              (операции анулирования переводов)
              удаляем референс проводки анулирования и
@@ -776,7 +780,8 @@ BEGIN
               lpad(nvl(ctkod_ku_1, '0'), 3, '0')||
               lpad(nvl(ctkod_k040_2, '0'), 3, '0')||
               lpad(nvl(ctkod_ku_2, '0'), 3, '0'),
-              ctkod_t071, nbuc1_, 'З XML по ШК'
+              ctkod_t071, ctkod_ku_1, 'З XML по ШК'
+--              ctkod_t071, nbuc1_, 'З XML по ШК'
           from NBUR_TMP_E9_SK
           where report_date = dat_rep_ and
                 kf = mfo_ and
@@ -790,7 +795,8 @@ BEGIN
               lpad(nvl(ctkod_ku_1, '0'), 3, '0')||
               lpad(nvl(ctkod_k040_2, '0'), 3, '0')||
               lpad(nvl(ctkod_ku_2, '0'), 3, '0'),
-              ctkod_t080, nbuc1_, 'З XML по ШК'
+              ctkod_t080, ctkod_ku_1, 'З XML по ШК'
+--              ctkod_t080, nbuc1_, 'З XML по ШК'
           from NBUR_TMP_E9_SK
           where report_date = dat_rep_ and
                 kf = mfo_ and
@@ -802,13 +808,17 @@ BEGIN
       end;
    end if;
 ---------------------------------------------------
-   DELETE FROM tmp_nbu
-   WHERE kodf = kodf_ AND datf = dat_;
----------------------------------------------------
-   INSERT INTO tmp_nbu (kodp, datf, kodf, znap, nbuc)
-      SELECT kodp, dat_, kodf_, SUM(to_number(znap)), nbuc
-        FROM rnbu_trace
-      GROUP BY KODP,NBUC;
+   if type_ != 'X'  then
+
+        DELETE FROM tmp_nbu
+         WHERE kodf = kodf_ AND datf = dat_;
+
+        INSERT INTO tmp_nbu (kodp, datf, kodf, znap, nbuc)
+           SELECT kodp, dat_, kodf_, SUM(to_number(znap)), nbuc
+             FROM rnbu_trace
+            GROUP BY KODP,NBUC;
+
+   end if;
 ----------------------------------------
    logger.info ('P_FE9_SB: End ');
 END p_fe9_sb;
